persistent-state: Extend rocksdb metrics
Collect more rocksdb metrics for compaction, sstables, and memtable
flushes. Also, the previous mechanism to collect block-cache metrics,
from the `get_memory_usage_stats()` function, was incorrect as we
needed a reference to the `Cache` objects (and pass them to the
function) in order to get the value. Instead, we'll just query the
standard "get property" function for them, like we do for most other
metrics.

Release-Note-Core: Add and correct many of the reported rocksdb
  metrics.
Change-Id: I6dc0cf3872343fbeb1e27b23fd9c563a5d649abe