adapter: Lower prepared_query_cache capacity
Each client connection will have a `prepared_query_cache`. The current
default capacity is hard coded to 10,000 elements. As this cache is only
used in the prepare/execute path, there's no need to have such a high
limit. This CL reduces it by an order of magnitude, down to 1000
elements (which is arguably still high).

Change-Id: Ib6128c549b812da0b6fa4e429b3f617c2729f49a