benchmarks: Delete memcached support
A transitive dependency (idna 0.2.3) of the memcached client crate has a
security vulnerability, and an intermediate dependency thereof looks
mostly abandoned and is on an old crate version, and another
intermediate dependency is still working on updating to the latest
version.  We don't really need this code, so let's delete it.

Change-Id: I44bf9058749c8dbfb43d9644a80a89db464145e7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8447
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>