readyset-server/controller: Select join sides for scan and lookup
Fix issue with joining on textual vs numerical or temporal columns.
We always use the left side column to scan and the right side column
to do the lookup, hence when the right side column is text, then
the non-textual column will be converted into text in order to
perform the lookup, which might not be properly matched.
Concider textual values: '100', '0100', '000100', which all
should match numerical value 100, but when we convert number
100 to text, we will get '100', so it will find only one match.
But if we select the textual column to scan, and the numerical
one to do the lookup, the above example will find all the matches.

Fixes: READ-4974
Change-Id: I5f956c9692e9d859de38fb783da749a02c1d191c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8454
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>