query_logger: Remove per-query map
We discovered a leak with the `QueryLogger::per_query_metrics`
field. This occurred when we received queries with embedded literals
as we used that string directly in the `SqlQuery` key to the
map. Hence, we grew the map in an unbounded way.

Use of that map was mostly eliminated in Icf9cbf6e, but one use was
remaining. This CL removes the map, and preserves that one metric that
was still stored in the values (whihch was a histoigram for query
parsing time).

Change-Id: I03b35f940c60b4433350cd0bb2f66e0acc22e927