adapter: ViewsSynchronizer batching
`ViewsSynchronizer` currently sends every pending query from the
`QueryStatusCache` to the controller's `/view_names` endpoint, to
discover the status of each query. In the (degenerate) case where
every query offered to the `QueryStatusCache` is unique (i.e. literals
in the sql string that we do not rewrite), we send the entire set of
pending queries to the controller. By default we set the upper bound
of `QueryStatusCache` to 100,000 elements. Thus, on a single pass of
`ViewsSynchronizer` we could ask the controller about 100,000
queries. This has several complications: the `ViewsSynchronizer` times
out before getting a response from the controller, and thus will just
end up submitting all the 100,000 elements again, while the controller
processes all 100,000 queries, only to have the work thrown away as the
caller disconnected as the processing times approach 60
seconds(!). The worst outcome of this is that no progress is actually
made, yet a lot of work is done and thrown away.

This CL simply sends small batches of queries to the controller, to
reduce the time required from the controller to process the queries,
and thus make progress on the overall count of queries in
`QueryStatusCache`. A counter is added as well to track the number of
queries that have been sent to the controller.

Running this locally, I found that the processing times come down to
less than 100 milliseconds, and the memory consumed was cut by 50% and
much more stable.

Change-Id: Iff853041a8997d623b12cf0eea222ebb005a05d3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8485
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>