alloc: Expose per-thread allocation stats in metrics
Note that we actually will have multiple threads with the same name
(e.g. tokio threads), but we can't use the thread ID (turning it into a
`u64` is behind an unstable feature flag), so they get collected
together into one counter for all threads with the same name. This
mainly means that multiple adapter threads will be aggregated, and at
our current stage I don't think we would get anything meaningful out of
breaking that up anyhow (i.e. we couldn't identify which tokio task was
responsible anyway). So this mainly enables separating out domains to be
able to identify if the problem is in dataflow, since domains do have a
1-1 thread-name correspondence.

Also note that although we drop the counters for threads that are no
longer reported, the metrics crate has no facility for "unregistering"
metrics, so those counters will be reset to 0 and then continuously
reported. This will likely lead to clutter as multiple compaction
threads, specifically, will continue to be reported after they have
ended.

Change-Id: Idf7a08b18fb85681e368c07c3d94ef5d4f4ac0bf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8418
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI