query_logger: Optimize QueryLogger
As `QueryLogger` is the consumer end of a
multi-producer, single-consumer channel. this CL makes some
optimizations to reduce processing time for each individual event.

Calculating the `QueryId` is now merged with getting the SQL string,
and done once; this avoids rewriting the string and anonymizing it
multiple times.

Performance optimizations are mainly on the non-verbose mode path,
where gains of 30-40% lower processing times can be observed using the
new benchmark that was added.

Change-Id: Ia2190517384676a7f6faee3bfca4acd37c894605
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8491
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>