readyset-adapter: Duplicate server's jemalloc profiling endpoints
Not seeing a worthwhile way to deduplicate this other than putting http
handling in the readyset-alloc crate which doesn't seem ideal, so this
is just copy and pasted.

As explained in the comment, having this in both places (port 6033 and
6034, effectively) might give the false impression that you can toggle
profiling on or off just for the adapter and not the server, and vice
versa, but that is not the case in standalone deployments where they are
in the same process and share a jemalloc instance.

Despite the duplication and the risk of confusion, I think it's worth
having it in both places so that one day when we have a distributed
deployment, we aren't asking ourselves why we need a restart to turn on
profiling in the adapter but not the server.

Change-Id: I89ecf1b284893148f67933cd3f12acbccdbf983d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8504
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>