readyset-util: Compact indexmaps less aggressively
After further reflection, a 10% limit on unused capacity seems way too
small.  If we delete a key soon after growing, we might needlessly copy
excessively.  Bumping to 2 isn't enough since the map grows by the same
factor, and we might copy excessively if the size hovers in the vicinity
of a power of two.  So we bump the garbage factor to 4 for purposes of
hysteresis.

Change-Id: Ia2443d094bbf2a96c67f59f1c223179a65d3cb7a