readyset-server: Evict much more often, less aggressively
Data from the field revealed some unacceptably large upquery times
correlated with evictions.  In a prior revision of eviction math, we
bumped up the indicated eviction sizes a bit in an attempt to stay
diligently within the proscribed memory limit.  This had the unwanted
side effect of making evictions larger and less frequent, which means
when they do happen, the domain blocks for longer, which is not what we
want.

Instead, we evict only the amount indicated by our heuristic
calculation, but we're lowering the period by a factor of ten (to 100
ms).  The higher frequency will spread out the evictions more evenly in
time so that the individual disruptions to upqueries will be smaller.

Release-Note-Core: Evictions are now smaller but more frequent to avoid
  spikes in upquery latency.
Change-Id: I7851daf69f59c11229d6c3f3ca67e2b830cde8d3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8498
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>