readyset-psql: Add CASCADE to drops in tests for more reliability
One of these tests failed while setting itself up with an error saying
the table couldn't be dropped due to a dependent object.  The test that
ran (serially) immediately before the failure created a view on its
table of the same name, so the problem seems to be not deleting the
table hard enough.

Change-Id: Ia72a7a391fb884229225a5a635186e4c6e90b26a