replicators: Start readyset from the minimum offset
During restart of the replicator, we need to find the correct position
to start streaming events. We have two concept of positions, the
schema position, which is updated after snapshot, and the per table
position, which is updated as each table gets events, or until we
update the min lower bound using ReplicationAction::LogPosition.
Our restart happens in three steps, first we attempt to snapshot new
tables, then we catch up with events between min applied position and
the current upstream position, then we start streaming events from the
current upstream position.
In order to find the min position, we were using
ReplicationOffset::max_offset, which has a clash between its name
saying it returns the max offset and documentation saying it returns
the min offset. The function is actually returning the max offset.

When starting readyset, we were getting the max_offset, then running
snapshot, which was updating the schema offset to current position,
and then running catch-up. In case of an error during catch-up, we
will restart the replicator, which causes us to fetch the max_offset
again, and this time it will be equal to upstream current position,
causing the catch-up to be skipped on the second time.

This patch changes the logic to properly use the min_offset when
starting replicator.

Fixes: REA-5098

Change-Id: I79b28005ce53964508319222a2f45807a9cd0ac0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8541
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>