readyset-common: Add LenMetric wrapper type for auto ..._len metrics
Wrap a T that is Len in this new type to automatically create a length
metric.  The value of the metric is updated on deref (but at the
beginning of the deref, not the end, but that tiny bit of slop should be
fine).

To create, you can call:

- new with metric name (if T: Default)
- new_meta with metric name and K/V metadata (if T: Default)
- new_from if collection creation requires parameters and T: From<...>

Those options should cover most common cases, but we can add more should
the need arise.

Change-Id: I8426fda5de6177a9cbbdb61e46600b282838c4d7