build: Parallelize 'run-tests'
Because nextest does not support running doctests, we are parallelizing
the run-tests.sh run in buildkite to run nextest in one job and doctests
in another.  This means spending a little more on CPU, but shortening
the total wall clock duration of a build (plus gaining the features
provided by nextest).

Change-Id: I8211a10ba41179d2b0ed1d888588863011df5afc