test: Cull some time zones to avoid test failures
We get recurring, random test failures when generating timestamps from
certain zones, mostly due to the fact that different databases use
different `tzdata` versions and some time zones are particularly
troublesome (i.e. America/Mexico_City, which was still on Lunar Mean
Time in 1922).  Without mapping tzdata used for our tests with the
tzdata used by supporting databases, we can't hope to verify proper
handling of *all* timezones for *all* edge cases.  This update removes
time zones which have been known to cause test failures in the past
and/or are unlikely to be encountered in the wild by users at this time.

Change-Id: Id46b3fadb2d8d9bfaecb8013ca0f69a25ade867f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8548
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI