migration: Send IsReady to all upstream full mat nodes
On restart, we have a race condition that causes a panic when
"recovering" the `DfState`. This occurs on joins with fully
materialized nodes, when replay begins. Before replay, all nodes will
be sent a `Ready` message, but base nodes (which must open files from
disk), "ready" themselves asynchronously - meaning, the Leader does
not block for them to become ready.

The problem we are encountering is when a replay that contains a join
node begins, the `DomainMigrationPlan` ensures only the left-side is
ready (by sending an `IsReady` message). The right-side has been sent
a `Ready` message, but we have not guaranteed that it is, indeed,
"ready".

This CL checks each ReplayPath, in the `DomainMigrationPlan`'s `Plan`
phase, and if the path contains fully-materialized nodes, it checks to
see if there's a join node. If there is a join, it walks up the DAG to
look for both base nodes and fully materialized nodes. The indices of
those nodes are added to a set, and a further part of the
`DomainMigrationPlan::commit()` will send an `IsReady` message to
those additional nodes.

This will not cause any additional traffic at startup as there's
already a `HashSet` in `DomainMigrationPlan::commit()` which tracks if
a `IsReady` has already been sent to a target node.

Fixes: REA-5126
Change-Id: Icbc815ce0a4aa4ed1b03aa734c785122c0839cbb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8563
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>