sql-passes: Respect aliases in JOIN clauses when resolving schemas
This is especially problematic when shadowing another table: we will
resolve the schema for the alias to the schema of the shadowed table.
For example, `select shadow.x from t1 join t2 as shadow` will end up
looking like `select public.shadow.x`. Since aliases are handled by a
later SQL rewrite pass, all we need to do (as we already do with the
more common `FROM <table> as <alias>` form of aliasing) is ignore it
when resolving the schema.

Fixes: REA-5188
Change-Id: Ib7bddd611aca99a04d9a024050f629dc34061999
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8584
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>