psql-srv: add PsqlValue::TinyText variant
When converting from DfValue to PsqlValue, we currently change
TinyText impls to Text. The problem here is Text uses a ThinArc, which
adds an unnecessary performance hindrance when no other actor would
be using the DfValue. I could see accessing/modifying the atomic
variable in the Arc showing a significant performance hit.

This CL introduces a PsqlValue::TinyText enum variant to avoid that
conversion to a Text, and thus avoiding the unnecessary Arc.

Addresses: REA-4984
Change-Id: I6cd0f439aeb2ca954d59a540adcccb1ea357fbff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8661
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>