psql-srv: Increase outbound buffer size
The existing `channel` uses a `Framed` instance, and we use the
default outbound (write) buffer size, which is set in tokio to 8
KiB. For most result set sizes we return that is completely
fine. However, on large (or very large) result sets, we can spend a
lot of time flushing to the socket due to the 8 KiB size.

This patch raises that buffer size to 64 KiB. On smaller responses,
this will have no impact. For larger responses (~400,000 in a result
set), the change reduces latencies by ~10% and increased throughput by
~8%.

Addresses: REA-4984
Change-Id: I911562f90c9d3763a7d9196ce64a3fc5eeaffd06