psql-src: Eliminate hidden clone in encode
When serializing Readyset-cached rows/columns to the output buffer,
there is a hidden clone in `psql-srv::Encoder`, which is causing a
perfomance hit. This CL replaces the existing iterator over the
columns to a loop which gets a reference to the column, and then
passes a reference to `put_binary_value()`/`put_text_value()`.

Addresses: REA-4984
Change-Id: I0e4c1bffb95bf9e281cee0abbb09c7b607ec96b4