psql-srv: Optimize Resultset::poll_next
When polling for the next result from a `ResultsetInner::Stream` in
`readyset_psql::resultset::ResultSet`, we call `collect()` after
mapping on the result set's columns.

The `collect()` functionality is taking up a lot of processing time
when looking at a set of flamegraphs. Since we already know the output
size of the vector in this path, this CL creates an correctly-sized
vec at the start, and replaces the `iter()` and `map()` constructs
with a simple for loop.

Further the other function constructs (iterate, enumerate) are also
slowing down processing. They are replaced, as well.

Currently, we create an instance of `PsqlValue`, and then `push()` it
into a vec. The push requires copying the `PsqlValue`. Thus, this CL
directly allocates the `PsqlValue` into the vec. This technique
requires the use of unsafe, but since this is the super hot path, it
is reasonable (and well as being very narrowly used).

Addresses: REA-4984
Change-Id: If10371ddeebaa0a85c3ea185ac33a9b7ed8b12cb