sql-passes: Fix schema resolution alias issue
We should not resolve the schema for table references which are
referring to aliases. We were incorrectly tracking the list of visible
aliases when walking select statements. A previous partial fix addressed
shadowing of non-replicated tables (REA-5188), but left the issue of
self-shadowing (REA-5139) unaddressed. While investigating, I realized
this would also lead to incorrect aliasing (thus skipping schema
resolution) in subqueries as table factors, which should not see aliases
on the same "level". This should be a more complete solution, with some
comments, and tests for these problems.

Fixes: REA-5139
Relates to: REA-5188

Change-Id: I67a788f80925bd2d075d9a0443b44835d90b168d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8603
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>