readyset-dataflow: Unquery keys evicted from readers from upstream nodes
When it's time to evict, readers have the best idea as to what's going
on because they track key usage using an approximately-LRU mechanism.
When a reader, in its wisdom, decides to evict a key for lack of use,
the rest of the system should should harken to its verdict.  If a key
isn't worth keeping in the reader, it's also not worth keeping in other
materializations (which have heretofore only ever reduced their size by
evicting keys at random).

Domains now implement this strategy.  After an eviction, the domain will
go through the evicted keys and send unqueries upward along the replay
path to the top of the dataflow graph.  At the top, we turn around and
start a normal eviction of the specified keys.

Change-Id: I7b98f288ff735cbb7a5634e45a95c74dd151bfd6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8679
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>