readyset-server: evict_single body should be optional
EvictSingle API accepts either a single key in the body or no body at
all. If no body is provided, we evict a random key from any partial
materialization.
The problem is that the bincode::deserialize does not handle empty
bodies, so we need to check if the body is empty and then deserialize
it if it is not.

Change-Id: I142abcd7888966da77f126fa6972d6af5229dcdb