mysql-srv: Merge PacketReader and PacketWriter into PacketConn
Previously, mysql-srv split the TcpStream into two streams to
create PacketReader and PacketWriter. However, this approach
is incompatible with TLS streams, as splitting a TLS stream is
not recommended (https://github.com/tokio-rs/tls/issues/40).

This commit merges PacketReader and PacketWriter into a single
PacketConn struct, which implements both the AsyncRead and
AsyncWrite traits.

This refactor lays the groundwork for MySQL TLS support.

Change-Id: I41ec4802ef3f7d0e423c3803b45376047bf64912
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8688
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>