readyset-clustertest: Add flag for manually skipping build
I noticed that the clustertest dependency build was making doctests
extremely slow, and (unlike clippy) there doesn't seem to be a way to
check that we are currently compiling doctests. So I added a flag to skip
building the readyset binary.

Additionally, I don't think these actually need to be release mode, so we
can further speed up build times that way.

Change-Id: I0bf44a9cbbfe291358bc37e667511bef650f3091