adapter: Convert StatementId to enum
`readyset-adapter::backend::StatementId` is currently a type alias
for `u32`, and currently represents the id of a statement prepared
against an upstream database. This CL changes it from a simple type
alias to an enum with two variants: one to represent the current
`u32`, and another to support "unnamed" prepared statements. The
definition is also moved into the `readyset-adapter-types` crate, so
it can be referenced easily. The "unnamed" behavior will be
implemented in followup patches in this CL list.

Change-Id: I1bb059077e6e5ce352ecddef05f242b8bb2004b7