adapter: Add UpstreamDatabase::query_with_params()
Add a function that allows us to call
`tokio_postgres::Client::query_typed_raw()`. This is the client
function that performs unnamed prepared statement calls to the
upstream postgres. This function returns it's own custom stream
type (`RowStream`), so we must percolate that up through the psql
stack.

MySQL does not support unnamed prepared statements, so that path
returns an appropriate error.

Change-Id: Ib57ef01e0d480f7d754511ac19d2532a554586ab