adapter: optimize PG unnamed prepared statements
This CL fixes our support for postgres's unnamed prepared statements
[0]. As unnamed prepared statements are very similar to regular, named
prepared statements, we already "supported" them (as the Frontend
messages are all the same). The problem was we were very inefficient -
even if the query was cached, we still made _two_ calls to the
upstream database: one to prepare the statement (as we treat it just
like a regular statement, that will be reused by the same name), and
one to close any prior unnamed statement we may have had
prepared. Thus, for something that could have been in our wheelhouse,
we were out to lunch.

This CL optimizes the support for unnamed prepared statements. We
still prepare the statement, but we only do it the first time we see
that query during the session. Note that we still need to prepare the
statment with the upstream as we need the type metadata for the
query (the types of incoming bind parameters, and the outgoing columns
in the result set, if any).

[0] https://www.postgresql.org/docs/current/protocol-flow.html

Fixes: REA-5238
Change-Id: I1b7f3ebef4382c74fcf5befb2559fcb57d01d317
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8771
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI