adapter: Cache unnamed prepared statements
Add a `SharedCache` to `readyset-adapter::BackendState` to maintain
query type information (params types and row types; input and output,
basically) for unnamed prepared statements. Using a `SharedCache`
allows us to share that type information across all connections, so we
only need to look it up (prepare the statement on the upstream) once
in the lifetime of the readyset process.

Change-Id: I953965ffdf286b1c330ba8de0d9baf5514153e8a