replicators: Refactor table filter code
This code took some effort to understand.

Before, the state in the TableFilter was split into (1) an allowlist
containing a map of schemas to either an allowlist or a denylist of
tables and (2) a denylist containing a map of schemas to either an
allowlist or a denylist of tables.  Yes, you probably read that
correctly.

The root allowlist and denylist were called explicitly_replicated and
replication_denied, respectively.  Either of these could hold another
allowlist or denylist per schema called ReplicateTableSpec::Tables or
ReplicateTableSpec::AllTablesExcept, respectively.

In order to make sense of what to do with this state, you had to
consider both of these 2-tiered lists at the same time, understand
certain conventions about what it meant when these collections were
empty, and be careful not to get them into invalid states when
performing modifications.

Now, we simplify the state to single map.  The keys are schemas and the
values are either an allowlist or denylist of tables for that schema.

To figure out if a table should be replicated, you look up the
corresponding list for its schema and see if the list allows the table
or not.

The old convention about empty collections is replaced with a boolean
that tells us what to do if we encounter a schema not registered with
the TableFilter.

During the refactor, I caught a bug via a test where a default schema
specified in the --upstream-db-url would no longer be respected if the
user also specified --replication-tables-ignore to ignore some of its
tables.  We would erroneously start replicating from other, unrelated
schemas.

This refactor fixes that bug, and the test is now updated to check for
the desired behavior.

Fixes: REA-5270
Change-Id: If35e7f85f47c0e418c32adc4beb20bf6b53d1838