replicators: Allow --replication-tables-ignore '*.*'
Starting Readyset with --replication-tables-ignore '*.*' will now deny
all tables.  This is intended to be useful for deployments that want to
initially start Readyset without replicating anything.

If desired, tables can still be added to a Readyset process running in
this mode via the ALTER READYSET ADD TABLES command.

Addresses: REA-5257
Change-Id: If2502611902a1e9685dbd58475f386ff66d79938
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8727
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>