fix: expanding * when subquery has non-col fields
Given a subquery that contains an expression that is not a
direct col ref (a + 2) or (max(a)) for example. The visitor
skips the column when reporting the schema to the outer query.
This patch allows the visitor to force subquery projections
to have an alias that it can report, else it returns an error.

Fixes: REA-4217
Change-Id: If6e07d11ae9e3080ebd89a3a2e06d62767ff8f0f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8766
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI