benchmarks: Remove dead config/functionality
There's a lot of excess reporting and execution modes that are
unused. One of the items removed here was a thing that reached out to
"noria" instances, scraped their premoethues metrics point(s), and
republished those metrics to a prometheus pushgateway. There was even
an entire module (src/utils/prometheus/) for parsing the prometheus
text and recreating its own metrics objects (counter, gauge,
etc). There is no need to re-export the metrics from another process,
and we can just use standard prometheus metrics for the things
reported by this benchmark app (which will be introduced in another
patch).

Change-Id: Ia759c116ea6e18752e84ce770f08b24d569ac231