readyset-sql: Convert ALTER TABLE
We are missing a lot of stuff that seems like it might affect us (such
as `MODIFY COLUMN`, which could change a column in a way where we would
no longer support it and would have to drop caches. It doesn't look like
we have been too bothered by that so far.

Change-Id: I5da416f28b40c184a698dcf6132df90deadbeef1