readyset-client: Parse MySQL DDL with readyset-sql-parsing
The `ChangeList::from_str` routine got in our way and has been replaced
with `ChangeList::from_strings`. It was only used for MySQL replication
DDL (which only feeds one statement at a time and probably doesn't
really need to go through this method, and should do what Postgres does)
and in tests (where it is mostly just a convenience for sending multiple
statements to `extend_recipe`).

We totally puke on `;` in nom-sql with the previous implementation (e.g.
we could split a query which included an escaped `;` in a table name,
and then fail to parse), which may be fine if it was just for tests.
Since it's not, let's make the tests more explicit and remove that
hazard.

This, then, lets us update to also parse with sqlparser-rs without
needing to do some other alternative splitting nonsense (or
re-implementing sqlparser's `parse_statements` method in a way that
supports our custom SQL parsing). One SQL statement at a time should be
enough for anyone.

Change-Id: I26d108636257ab652915eb6573251c0deadbeef1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8822
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>