readyset-psql: Updates to driver unnamed pStatements
Pull in changes to `tokio-postgres` driver, which affect the unnamed
prepared statement API, `client::query_typed_raw()`.

The inputs to the function now take a result transfer format, which
allows us to pass in the format that readyset was asked for from its
client. Without this we could return data in a format the client was
not expecting (in particular, pgjdbc requests `Text` result format, yet
we were always returning `Binary`).

The output type of the function still returns `RowStream`, but it's
iterator's type is now `GenericResult`. `GenericResult` can emit
either a `Row` or a `Command`; the prior `RowStream` emitted type was
only `Row`. We need both types, similar to how we process regular
prepared statements, in case the client sends commands like `BEGIN` as
a prepared statement (either named or unnamed). pgjdbc, in particular,
actually checks the response message to the command and will fail if
not what it should be; anecdotally, most other drivers don't bother
checking the response for the command.

Change-Id: If4c411004a3d13783b71363689d6f0fa69378858