adapter: Pass correct Types to tokio_postgres
When support for postgres unnamed prepared statements was added, on
`execute()` we derived the postgres types from the `DfValue`s that
were bound into the query. Reusing the existing function to go
from `DfValue` to `DfType` to postgres `Type` were optimistic as best,
but incorrect for many types. There's fortunately a much more
bullet-proof solution, and we've already calculated it by this time.

When we prepared the statement, we already get the `Type`s, we just
need to keep it around in `readyset-psql::Upstream`. Thus, this CL
pushes handling of the unnamed prepared statement down into the
`readyset-psql::Upstream`, which allows us to better maintain the
`Type` information, and it also cleans up some of the messiness that
bled into the `adapter::Backend`.

Change-Id: I51bb6f6bca184ccb82bb348ec3831ea83eb727e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8808
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI