readyset-sql: Fix variable scope conversion
I'm not completely sure, but I think we sometimes go down one variable
name conversion path rather than the other based on whether we are in a
`SET` or `SELECT` statement. Both branches should work correctly now.

And I don't know where I got the idea that session was the default
instead of local, but that's fixed too.

Change-Id: Ic44ffdce1c3636f37bde08d6144166adeadbeef2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8826
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI