readyset-sql-parsing: Parse ALTER READYSET statements
This is just RESNAPSHOT TABLE and ADD TABLES.

I added a `parse_readyset_keywords` function to parse all keywords at
once, and added a variant of `ReadysetKeyword` to allow intermixing of
custom Readyset keywords with native sqlparser keywords. This allows
parsing a list of them atomically, which may avoid shadowing correct
parses that we would get by falling back to sqlparser, whereas if we
error out or fall back having parsed one keyword but not matching the
second, we could get incorrect parses.

Change-Id: I71276f4db84403877df386892b0f5a11deadbeef