mysql-srv: Support for caching_sha2_password
This commit adds support for the caching_sha2_password
authentication method. Previously, the implementation only
supported mysql_native_password, which is no longer
available in MySQL 9.0.

All authentication paths related to caching_sha2_password
are now implemented, including FastAuth, FullAuth with TLS,
and FullAuth with a public key request.

Additionally, this commit addresses issues in the C API for
mysql_change_user function, where the password was not sent
in the initial request. The updated implementation follows
the same flow as the initial handshake, sending a switch
authentication packet and requesting the password.

Change-Id: I76db7b865163e7e12b02e3c712a7a539334054e2