Allow multiple users to connect to Readyset
This commit allows multiple users to connect to Readyset by
adding a new parameter --allowed-users containing a comma-separated
list of username:password that will be used to authenticate
connections to the Readyset server. The same user and password will
be used for both the Readyset server and the upstream database.

This commit also remove the obsolete --username and --password
parameters.

Closes: REA-4641
Closes: #1349

Release-Note-Core: It's now possible to connect to Readyset with
  multiple users by specifying a comma-separated list of
  username:password pairs in the --allowed-users parameter.
Change-Id: Ie2e44c68a386e164278e95edfc76a67b6406cf57