buildkite: Remove benchmarks test
This test currently holds no value as we are not actively using the
`//benchmarks` crate, especially in the way that is was being tested
here (in "local" mode). Later CLs in this chain will remove some of
the functionality referenced in this test, so remove it now. We may
choose to re-introduce some buildtime CI test for benchmarks, but it
will be different.

Change-Id: Idca24a603f8532ec8c63b129d9150a68bfc87c68