builtin function: `json[b]_object_agg`
Add basic support of json_object_agg. Current implemention of
aggregator only allows agg over a single col. We implement
a workaround where we inject a call to `json_build_object` over the
key and value columns and hard-code an alias for that call
and give that alias to the aggregator. There are some potential
performance issues that should definitely be addressed.

Release-Note-Core: Added support for `json[b]_object_agg` and
`json_objectagg`.
Fixes: REA-4369
Change-Id: I5c5548e0a491b1ba1adf49c0699f2308c8ab0acf

Change-Id: Id217ab818cf545f0eac8bebe6e41ba98ce91b999