builtin function: `json[b]_object_agg`
Add basic support of json_object_agg. Current implementation of the
aggregator only allows agg over a single column. We implement
a workaround where we inject a call to `json_build_object` over the
key and value columns, hard-code an alias for that call,
and give that alias to the aggregator.

Release-Note-Core: Added support for builtin functions
  `json[b]_object_agg` and `json_objectagg`.
Fixes: REA-4369
Change-Id: I5c5548e0a491b1ba1adf49c0699f2308c8ab0acf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8848
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI