readyset-sql: Change ordering of CREATE TABLE column options in conversion
There is literally no good reason for this to be one order versus the
other, it just matches existing tests this way. If we cared about
preserving the input ordering, at some point we could upstream a fix to
sqlparser-rs which treats table options as a Vec (like column options)
instead of individual fields. But I think we only care about matching
the input for proxied queries where users may be trying to find a
specific query; DDL we don't generally show back to the user, and we
only care about some options and not their order.

Change-Id: Ic0130359d97255303d1d5d293806d2aec7a8b7c9