debug: Update to sqlparser fork with latest fixes
Importantly, this version of sqlparser supports `COLLATE =` syntax,
which some framework tests need, and a few of my fixes which have yet to
land in upstream.

The idea is to replace this patch, in the future, with one that uses an
appropriately pinned commit in the official sqlparser repository. It
will be WIP until then.

Change-Id: Ib12021831913a2590b016a392d7597c7a84d541a