psql-srv: Appropriately return EmptyQueryResponse message
When a PostgreSQL client sends an empty query over the wire protocol,
the appropriate response is to send an EmptyQueryResponse message in
reply.  This is different from the CommandComplete message, which is the
standard reply for regular queries.

Our usage of our fork of tokio_postgres elides a copy during standard
query processing by directly accessing the CommandComplete message
buffer, which is populated by tokio_postgres when Readyset runs queries
on an upstream PostgreSQL database.  Unfortunately, tokio_postgres
returns a CommandComplete message for an empty query run on the
upstream, even though the upstream replied with an EmptyQueryResponse
message.  Fortunately, however, we don't have to forward all
CommandComplete messages we process from tokio_postgres back to our
client.

Now, we selectively detect when tokio_postgres produces a
CommandComplete message that is semantically equivalent to an
EmptyQueryResponse, and then we send back an EmptyQueryResponse message
to our client.

This patch brings our behavior a little more into alignment with the
PostgreSQL wire protocol, allowing us to be a little more compatible
with applications that expect to receive an EmptyQueryResponse in reply
to an empty query.

Also add some flushes to related tests, as I experienced some flakiness
there with the stream not being processed before the test functions
exited.

Fixes: REA-5385
Release-Note-Core: Improve PostgreSQL wire protocol compatibility by
  appropriately responding to an empty query with an EmptyQueryResponse
  message.
Change-Id: Ib24aa57f9f38663946feefee52e6c862d1f9daf3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8898
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI